requirements.txt
fastapi
uvicorn
telethon
httpx
apscheduler
python-dotenv
aiosqlite
